+++ src/System/Taffybar/Widget.hs
@@ -12,9 +12,6 @@ module System.Taffybar.Widget
   -- * "System.Taffybar.Widget.Crypto"
   , module System.Taffybar.Widget.Crypto
 
-  -- * "System.Taffybar.Widget.Decorators"
-  , module System.Taffybar.Widget.Decorators
-
   -- * "System.Taffybar.Widget.DiskIOMonitor"
   , module System.Taffybar.Widget.DiskIOMonitor
 
@@ -68,7 +65,6 @@ import System.Taffybar.Widget.Battery
 import System.Taffybar.Widget.CPUMonitor
 import System.Taffybar.Widget.CommandRunner
 import System.Taffybar.Widget.Crypto
-import System.Taffybar.Widget.Decorators
 import System.Taffybar.Widget.DiskIOMonitor
 import System.Taffybar.Widget.FSMonitor
 import System.Taffybar.Widget.FreedesktopNotifications